Chore: 스토리북 제거
