+++ package.json
@@ -15,8 +15,6 @@
     "format:fix": "pnpm run prettier:fix && pnpm run eslint:fix",
     "preview": "vite preview",
     "prepare": "husky",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build",
     "lintStaged": "lint-staged",
     "checkTs": "tsc --noEmit"
   },
@@ -37,16 +35,6 @@
   "devDependencies": {
     "@commitlint/cli": "^19.1.0",
     "@commitlint/config-conventional": "^19.1.0",
-    "@storybook/addon-essentials": "^7.6.15",
-    "@storybook/addon-interactions": "^7.6.15",
-    "@storybook/addon-links": "^7.6.15",
-    "@storybook/addon-onboarding": "^1.0.11",
-    "@storybook/blocks": "^7.6.15",
-    "@storybook/jest": "^0.2.3",
-    "@storybook/react": "^7.6.15",
-    "@storybook/react-vite": "^7.6.15",
-    "@storybook/test": "^7.6.15",
-    "@storybook/testing-library": "^0.2.2",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
@@ -64,7 +52,6 @@
     "lint-staged": "^15.2.2",
     "msw": "^2.3.4",
     "prettier": "^3.2.5",
-    "storybook": "^7.6.15",
     "typescript": "^5.2.2",
     "vite": "^5.1.0",
     "vitest": "^1.2.2"
