+++ src/app/App.tsx
@@ -1,0 +1,16 @@
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import "./App.css";
+import { MainPage } from "@pages/main";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <MainPage />,
+  },
+]);
+
+function App() {
+  return <RouterProvider router={router} />;
+}
+
+export default App;
