+++ src/pages/main/ui/Page/Page.tsx
@@ -1,0 +1,3 @@
+export function MainPage() {
+  return <div>Home</div>;
+}
